requirements.txt
pdfkit==1.0.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
SQLAlchemy==2.0.31
